refactor(identity-agent): export TMPXSealer + BuildKeyStore, drop duplicates from reference
The reference identity-agent and the production targeting/identityagent
held parallel copies of the same TMPX and keystore primitives. After
exposing them on the library, the reference shrinks to its actual job:
flag parsing, audience-only wiring, and an inline handler.

Library API changes (targeting/identityagent)
  - Exported BuildKeyStore (was buildKeyStore). No behavior change beyond
    the signature being callable from outside the package.
  - Renamed the runtime TMPX type to TMPXSealer (was tmpxConfig) with a
    Seal(ids []tmproto.IdentityToken) (string, error) method that
    replaces the free buildTmpxToken function.
  - Constructor: NewTMPXSealer(ctx, TMPXConfig, *slog.Logger) replaces
    loadTmpxConfig. Same semantics — partial config rejects, missing
    stub-ack rejects, JWKS background refresh goroutine launches.
  - selectEntries is now a method on *TMPXSealer (was selectTmpxEntries).
  - IdentityHandlerConfig.TMPXConfig field renamed to TMPXSealer to
    match the new type name.

Reference cleanup
  - Deleted: buildKeyStore, loadTmpxConfig, parseTmpxPriority,
    uidToTmpxTypeID, buildTmpxToken, selectTmpxEntries, indexOfUIDType,
    stubBinaryToken, tmpxConfig, tmpxRecipientResolver from
    reference/identity-agent/cmd/identity-agent/main.go.
  - Imports: crypto/sha512, sort dropped; identityagent added.
  - Reference still owns its flag-resolution helpers (resolveAddr,
    resolveString, setFlags), the TMP_IDENTITY_* env names, and the
    TMP_IDENTITY_TMPX_REFERENCE_STUB_ACK legacy env (translated to
    TMPXConfig.ReferenceStubAck at the boundary).

Tests
  - reference/identity-agent/cmd/identity-agent/tmpx_test.go moved to
    targeting/identityagent/tmpx_test.go and rewritten against the new
    API. Same coverage; now exercises the canonical impl directly. The
    reference module no longer ships tests of its own (all primitives
    live in the library).
