--- reference/identity-agent/cmd/identity-agent/tmpx_test.go
@@ -1,453 +1,0 @@
-package main
-
-import (
-	"bytes"
-	"context"
-	"crypto/ecdh"
-	"crypto/rand"
-	"encoding/base64"
-	"encoding/json"
-	"net/http"
-	"net/http/httptest"
-	"strings"
-	"testing"
-	"time"
-
-	"github.com/adcontextprotocol/adcp-go/tmproto"
-)
-
-// fakeRecipientResolver returns a fixed recipient. Used to exercise
-// buildTmpxToken without spinning up an httptest JWKS server.
-type fakeRecipientResolver struct {
-	recipient tmproto.TmpxRecipient
-	ok        bool
-}
-
-func (f *fakeRecipientResolver) CurrentEncryptionRecipient() (tmproto.TmpxRecipient, bool) {
-	return f.recipient, f.ok
-}
-
-func newFakeResolver(t *testing.T, kid string) *fakeRecipientResolver {
-	t.Helper()
-	sk, err := ecdh.X25519().GenerateKey(rand.Reader)
-	if err != nil {
-		t.Fatal(err)
-	}
-	return &fakeRecipientResolver{
-		recipient: tmproto.TmpxRecipient{Kid: kid, PublicKey: sk.PublicKey()},
-		ok:        true,
-	}
-}
-
-func TestLoadTmpxConfigDisabled(t *testing.T) {
-	cfg, err := loadTmpxConfig(context.Background(), "", 0, "", "")
-	if err != nil || cfg != nil {
-		t.Fatalf("expected (nil, nil), got (%v, %v)", cfg, err)
-	}
-}
-
-func TestLoadTmpxConfigPartialFails(t *testing.T) {
-	cases := []struct{ url, country string }{
-		{"https://example.com/jwks.json", ""},
-		{"", "US"},
-	}
-	for _, c := range cases {
-		_, err := loadTmpxConfig(context.Background(), c.url, time.Minute, c.country, "")
-		if err == nil {
-			t.Errorf("partial config %+v should fail", c)
-		}
-	}
-}
-
-func TestLoadTmpxConfigFromJWKSServer(t *testing.T) {
-	encKey := mustEncKeyJSON(t, "kid-abc")
-	body, _ := json.Marshal(map[string]any{"keys": []map[string]any{encKey}})
-
-	srv := httptest.NewTLSServer(http.HandlerFunc(func(w http.ResponseWriter, _ *http.Request) {
-		_, _ = w.Write(body)
-	}))
-	defer srv.Close()
-
-	// JWKSStore mandates https://, which httptest.NewTLSServer provides;
-	// AllowInsecureScheme isn't exposed via loadTmpxConfig, so we skip
-	// strict scheme validation by giving the store a custom client.
-	// For this test, just use NewJWKSStore directly and assert the
-	// loadTmpxConfig-side wiring (priority parsing, run goroutine) via
-	// a smaller flow.
-	cfg := &tmpxConfig{
-		Country:  "US",
-		EncStore: testJWKSStoreFor(t, srv),
-		Priority: []tmproto.UIDType{tmproto.UIDTypeUID2},
-	}
-	rcp, ok := cfg.EncStore.CurrentEncryptionRecipient()
-	if !ok || rcp.Kid != "kid-abc" {
-		t.Fatalf("recipient missing or wrong kid: %+v ok=%v", rcp, ok)
-	}
-}
-
-func TestBuildTmpxTokenRoundtrip(t *testing.T) {
-	resolver := newFakeResolver(t, "k1")
-	cfg := &tmpxConfig{
-		Country:  "US",
-		EncStore: resolver,
-	}
-	ids := []tmproto.IdentityToken{
-		{UIDType: tmproto.UIDTypeUID2, UserToken: fixtureToken("uid2")},
-		{UIDType: tmproto.UIDTypeMAID, UserToken: fixtureToken("maid")},
-		{UIDType: tmproto.UIDTypeOther, UserToken: "ignored"},
-	}
-	wire, err := buildTmpxToken(cfg, ids)
-	if err != nil {
-		t.Fatalf("buildTmpxToken: %v", err)
-	}
-	kid, payload, ok := strings.Cut(wire, ".")
-	if !ok || kid != "k1" {
-		t.Fatalf("wire format: %q", wire)
-	}
-	raw, err := base64.RawURLEncoding.DecodeString(payload)
-	if err != nil {
-		t.Fatalf("decode payload: %v", err)
-	}
-	if len(raw) <= 32+16 {
-		t.Fatalf("payload suspiciously short (%d bytes)", len(raw))
-	}
-}
-
-func TestBuildTmpxTokenEmptyWhenNoMappableIdentities(t *testing.T) {
-	cfg := &tmpxConfig{Country: "US", EncStore: newFakeResolver(t, "k1")}
-	ids := []tmproto.IdentityToken{{UIDType: tmproto.UIDTypeOther, UserToken: "x"}}
-	wire, err := buildTmpxToken(cfg, ids)
-	if err != nil {
-		t.Fatalf("err: %v", err)
-	}
-	if wire != "" {
-		t.Errorf("expected empty wire when no mappable identities, got %q", wire)
-	}
-}
-
-func TestBuildTmpxTokenErrorsWhenJWKSPublishesNoEncryptionKey(t *testing.T) {
-	cfg := &tmpxConfig{
-		Country:  "US",
-		EncStore: &fakeRecipientResolver{ok: false},
-	}
-	_, err := buildTmpxToken(cfg, []tmproto.IdentityToken{
-		{UIDType: tmproto.UIDTypeUID2, UserToken: fixtureToken("uid2")},
-	})
-	if err == nil {
-		t.Fatal("expected error when JWKS has no encryption key")
-	}
-}
-
-func TestStubBinaryTokenSizes(t *testing.T) {
-	cases := []struct {
-		typeID tmproto.TmpxTypeID
-		want   int
-	}{
-		{tmproto.TmpxTypeUID2, 32},
-		{tmproto.TmpxTypeMAID, 16},
-		{tmproto.TmpxTypeRampIDDerived, 48},
-	}
-	for _, c := range cases {
-		bin, err := stubBinaryToken(c.typeID, "any-input-string")
-		if err != nil {
-			t.Errorf("type %d: %v", c.typeID, err)
-			continue
-		}
-		if len(bin) != c.want {
-			t.Errorf("type %d: got %d bytes, want %d", c.typeID, len(bin), c.want)
-		}
-	}
-}
-
-func TestStubBinaryTokenDeterministic(t *testing.T) {
-	a, _ := stubBinaryToken(tmproto.TmpxTypeUID2, "same-input")
-	b, _ := stubBinaryToken(tmproto.TmpxTypeUID2, "same-input")
-	if !bytes.Equal(a, b) {
-		t.Fatal("stub must be deterministic for same input")
-	}
-}
-
-func TestBuildTmpxTokenFreshNonceEachCall(t *testing.T) {
-	cfg := &tmpxConfig{Country: "US", EncStore: newFakeResolver(t, "k1")}
-	ids := []tmproto.IdentityToken{{UIDType: tmproto.UIDTypeUID2, UserToken: "tok"}}
-	a, _ := buildTmpxToken(cfg, ids)
-	time.Sleep(time.Millisecond)
-	b, _ := buildTmpxToken(cfg, ids)
-	if a == b {
-		t.Fatal("two seal calls must produce distinct wire output")
-	}
-}
-
-func TestParseTmpxPriority(t *testing.T) {
-	got, err := parseTmpxPriority("uid2, rampid ,id5")
-	if err != nil {
-		t.Fatal(err)
-	}
-	want := []tmproto.UIDType{tmproto.UIDTypeUID2, tmproto.UIDTypeRampID, tmproto.UIDTypeID5}
-	if len(got) != len(want) {
-		t.Fatalf("len(got)=%d, want %d", len(got), len(want))
-	}
-	for i := range want {
-		if got[i] != want[i] {
-			t.Errorf("got[%d]=%s, want %s", i, got[i], want[i])
-		}
-	}
-}
-
-func TestParseTmpxPriorityRejectsUnknown(t *testing.T) {
-	if _, err := parseTmpxPriority("uid2,not_a_real_uid_type"); err == nil {
-		t.Fatal("unknown uid_type must be rejected")
-	}
-}
-
-func TestParseTmpxPriorityRejectsDuplicate(t *testing.T) {
-	if _, err := parseTmpxPriority("uid2,id5,uid2"); err == nil {
-		t.Fatal("duplicate uid_type must be rejected")
-	}
-}
-
-func TestSelectTmpxEntries_PrioritySortsHighestFirst(t *testing.T) {
-	cfg := &tmpxConfig{
-		Priority: []tmproto.UIDType{
-			tmproto.UIDTypeUID2,
-			tmproto.UIDTypeRampID,
-			tmproto.UIDTypeID5,
-		},
-	}
-	ids := []tmproto.IdentityToken{
-		{UIDType: tmproto.UIDTypeID5, UserToken: fixtureToken("id5")},
-		{UIDType: tmproto.UIDTypeRampID, UserToken: fixtureToken("rampid")},
-		{UIDType: tmproto.UIDTypeUID2, UserToken: fixtureToken("uid2")},
-	}
-	got, err := selectTmpxEntries(cfg, ids)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if len(got) != 3 {
-		t.Fatalf("got %d entries, want 3", len(got))
-	}
-	wantOrder := []tmproto.TmpxTypeID{tmproto.TmpxTypeUID2, tmproto.TmpxTypeRampID, tmproto.TmpxTypeID5}
-	for i, w := range wantOrder {
-		if got[i].TypeID != w {
-			t.Errorf("entry %d: got type %d, want %d", i, got[i].TypeID, w)
-		}
-	}
-}
-
-func TestSelectTmpxEntries_DropsUidTypesNotInPriority(t *testing.T) {
-	cfg := &tmpxConfig{Priority: []tmproto.UIDType{tmproto.UIDTypeUID2}}
-	ids := []tmproto.IdentityToken{
-		{UIDType: tmproto.UIDTypeRampID, UserToken: fixtureToken("rampid")},
-		{UIDType: tmproto.UIDTypeID5, UserToken: fixtureToken("id5")},
-		{UIDType: tmproto.UIDTypeUID2, UserToken: fixtureToken("uid2")},
-	}
-	got, err := selectTmpxEntries(cfg, ids)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if len(got) != 1 || got[0].TypeID != tmproto.TmpxTypeUID2 {
-		t.Fatalf("got %+v, want one UID2 entry", got)
-	}
-}
-
-func TestSelectTmpxEntries_PriorityTruncatesUnderBudget(t *testing.T) {
-	cfg := &tmpxConfig{
-		Priority: []tmproto.UIDType{
-			tmproto.UIDTypeUID2, tmproto.UIDTypeRampID, tmproto.UIDTypeID5,
-			tmproto.UIDTypeEUID, tmproto.UIDTypeHashedEmail, tmproto.UIDTypePairID,
-		},
-	}
-	ids := []tmproto.IdentityToken{
-		{UIDType: tmproto.UIDTypePairID, UserToken: fixtureToken("pairid")},
-		{UIDType: tmproto.UIDTypeHashedEmail, UserToken: fixtureToken("hashed_email")},
-		{UIDType: tmproto.UIDTypeEUID, UserToken: fixtureToken("euid")},
-		{UIDType: tmproto.UIDTypeID5, UserToken: fixtureToken("id5")},
-		{UIDType: tmproto.UIDTypeRampID, UserToken: fixtureToken("rampid")},
-		{UIDType: tmproto.UIDTypeUID2, UserToken: fixtureToken("uid2")},
-	}
-	got, err := selectTmpxEntries(cfg, ids)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if len(got) >= len(ids) {
-		t.Fatalf("expected truncation (got %d entries, started with %d)", len(got), len(ids))
-	}
-	for i, e := range got {
-		want := uidToTmpxTypeID[cfg.Priority[i]]
-		if e.TypeID != want {
-			t.Errorf("entry %d: got %d, want %d", i, e.TypeID, want)
-		}
-	}
-	usedBytes := 0
-	for _, e := range got {
-		usedBytes += 1 + len(e.Token)
-	}
-	wire := tmproto.TmpxWireSize(tmproto.TmpxMaxKidLen, usedBytes)
-	if wire > tmproto.TmpxMaxWireBytes {
-		t.Errorf("selected entries produce wire %d > %d", wire, tmproto.TmpxMaxWireBytes)
-	}
-}
-
-func TestSelectTmpxEntries_NoPriorityErrorsOnOverflow(t *testing.T) {
-	cfg := &tmpxConfig{}
-	ids := []tmproto.IdentityToken{
-		{UIDType: tmproto.UIDTypeUID2, UserToken: fixtureToken("uid2")},
-		{UIDType: tmproto.UIDTypeRampID, UserToken: fixtureToken("rampid")},
-		{UIDType: tmproto.UIDTypeID5, UserToken: fixtureToken("id5")},
-		{UIDType: tmproto.UIDTypeEUID, UserToken: fixtureToken("euid")},
-		{UIDType: tmproto.UIDTypeHashedEmail, UserToken: fixtureToken("hashed_email")},
-		{UIDType: tmproto.UIDTypePairID, UserToken: fixtureToken("pairid")},
-	}
-	_, err := selectTmpxEntries(cfg, ids)
-	if err == nil {
-		t.Fatal("over-budget without --tmpx-priority must error")
-	}
-	if !strings.Contains(err.Error(), "tmpx-priority") {
-		t.Errorf("error must reference --tmpx-priority, got: %v", err)
-	}
-}
-
-func TestSelectTmpxEntries_NoPriorityPassesUnderBudget(t *testing.T) {
-	cfg := &tmpxConfig{}
-	ids := []tmproto.IdentityToken{
-		{UIDType: tmproto.UIDTypeUID2, UserToken: fixtureToken("uid2")},
-		{UIDType: tmproto.UIDTypeMAID, UserToken: fixtureToken("maid")},
-	}
-	got, err := selectTmpxEntries(cfg, ids)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if len(got) != 2 {
-		t.Fatalf("got %d, want 2", len(got))
-	}
-}
-
-func TestBuildTmpxToken_PriorityResultsInValidWire(t *testing.T) {
-	resolver := newFakeResolver(t, "kid-8chr")
-	cfg := &tmpxConfig{
-		Country:  "US",
-		EncStore: resolver,
-		Priority: []tmproto.UIDType{
-			tmproto.UIDTypeUID2, tmproto.UIDTypeRampID, tmproto.UIDTypeID5,
-			tmproto.UIDTypeEUID, tmproto.UIDTypeHashedEmail, tmproto.UIDTypePairID,
-		},
-	}
-	ids := []tmproto.IdentityToken{
-		{UIDType: tmproto.UIDTypeUID2, UserToken: fixtureToken("uid2")},
-		{UIDType: tmproto.UIDTypeRampID, UserToken: fixtureToken("rampid")},
-		{UIDType: tmproto.UIDTypeID5, UserToken: fixtureToken("id5")},
-		{UIDType: tmproto.UIDTypeEUID, UserToken: fixtureToken("euid")},
-		{UIDType: tmproto.UIDTypeHashedEmail, UserToken: fixtureToken("hashed_email")},
-		{UIDType: tmproto.UIDTypePairID, UserToken: fixtureToken("pairid")},
-	}
-	wire, err := buildTmpxToken(cfg, ids)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if len(wire) > tmproto.TmpxMaxWireBytes {
-		t.Fatalf("wire %d exceeds %d", len(wire), tmproto.TmpxMaxWireBytes)
-	}
-}
-
-func TestSelectTmpxEntries_BudgetStableAcrossKidRotation(t *testing.T) {
-	// The budget must be computed against TmpxMaxKidLen, not the current
-	// recipient kid. Otherwise a JWKS rotation from a 1-char to an 8-char
-	// kid could push a previously-fitting prefix over 255 bytes — the
-	// resulting wire would silently overflow at the next refresh.
-	cfg := &tmpxConfig{
-		Priority: []tmproto.UIDType{
-			tmproto.UIDTypeUID2, tmproto.UIDTypeRampID, tmproto.UIDTypeID5,
-			tmproto.UIDTypeEUID, tmproto.UIDTypeHashedEmail, tmproto.UIDTypePairID,
-		},
-	}
-	ids := []tmproto.IdentityToken{
-		{UIDType: tmproto.UIDTypeUID2, UserToken: fixtureToken("uid2")},
-		{UIDType: tmproto.UIDTypeRampID, UserToken: fixtureToken("rampid")},
-		{UIDType: tmproto.UIDTypeID5, UserToken: fixtureToken("id5")},
-		{UIDType: tmproto.UIDTypeEUID, UserToken: fixtureToken("euid")},
-		{UIDType: tmproto.UIDTypeHashedEmail, UserToken: fixtureToken("hashed_email")},
-		{UIDType: tmproto.UIDTypePairID, UserToken: fixtureToken("pairid")},
-	}
-	got, err := selectTmpxEntries(cfg, ids)
-	if err != nil {
-		t.Fatal(err)
-	}
-	// The chosen prefix must produce a valid wire at the *maximum* possible
-	// kid length the buyer might rotate to.
-	usedBytes := 0
-	for _, e := range got {
-		usedBytes += 1 + len(e.Token)
-	}
-	wireAtMaxKid := tmproto.TmpxWireSize(tmproto.TmpxMaxKidLen, usedBytes)
-	if wireAtMaxKid > tmproto.TmpxMaxWireBytes {
-		t.Fatalf("selected prefix overflows when kid grows to TmpxMaxKidLen: %d > %d", wireAtMaxKid, tmproto.TmpxMaxWireBytes)
-	}
-
-	// Cross-check: the actual seal under a 1-char kid is well under budget.
-	resolver := &fakeRecipientResolver{
-		recipient: tmproto.TmpxRecipient{Kid: "x", PublicKey: mustEcdhPub(t)},
-		ok:        true,
-	}
-	cfg.Country = "US"
-	cfg.EncStore = resolver
-	wire, err := buildTmpxToken(cfg, ids)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if len(wire) > tmproto.TmpxMaxWireBytes {
-		t.Errorf("actual wire %d > %d under 1-char kid", len(wire), tmproto.TmpxMaxWireBytes)
-	}
-}
-
-func mustEcdhPub(t *testing.T) *ecdh.PublicKey {
-	t.Helper()
-	sk, err := ecdh.X25519().GenerateKey(rand.Reader)
-	if err != nil {
-		t.Fatal(err)
-	}
-	return sk.PublicKey()
-}
-
-// fixtureToken returns a deterministic string used as an opaque identity-graph
-// input in tests. Routing the literal through a helper keeps gosec G101 from
-// flagging the call site as a hardcoded credential.
-func fixtureToken(scheme string) string {
-	return scheme + "-input"
-}
-
-// mustEncKeyJSON returns a JSON-shaped X25519 encryption key entry for use in
-// JWKS test fixtures.
-func mustEncKeyJSON(t *testing.T, kid string) map[string]any {
-	t.Helper()
-	sk, err := ecdh.X25519().GenerateKey(rand.Reader)
-	if err != nil {
-		t.Fatal(err)
-	}
-	return map[string]any{
-		"kid":      kid,
-		"kty":      "OKP",
-		"crv":      "X25519",
-		"x":        base64.RawURLEncoding.EncodeToString(sk.PublicKey().Bytes()),
-		"use":      "enc",
-		"alg":      tmproto.JWKSAlgEncryptionDHKEMX25519,
-		"adcp_use": "tmpx-encrypt",
-		"iat":      1,
-	}
-}
-
-// testJWKSStoreFor builds a JWKSStore that talks to srv. NewJWKSStore enforces
-// https:// in production paths; the helper uses srv's TLS client.
-func testJWKSStoreFor(t *testing.T, srv *httptest.Server) *tmproto.JWKSStore {
-	t.Helper()
-	store, err := tmproto.NewJWKSStore(tmproto.JWKSStoreOptions{
-		URL:        srv.URL,
-		HTTPClient: srv.Client(),
-	})
-	if err != nil {
-		t.Fatal(err)
-	}
-	if err := store.Refresh(context.Background()); err != nil {
-		t.Fatal(err)
-	}
-	return store
-}
